refactor: dto 내부에서 엔티티 속성 꺼내도록 변경
