--- src/main/java/roomescape/ReservationResponse.java
@@ -1,7 +1,0 @@
-package roomescape;
-
-import java.time.LocalDate;
-import java.time.LocalTime;
-
-public record ReservationResponse(Long id, String name, LocalDate date, LocalTime time) {
-}
