+++ src/main/java/roomescape/controller/dto/ReservationRequest.java
@@ -1,4 +1,4 @@
-package roomescape;
+package roomescape.controller.dto;
 
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
